Refactor frontend and enhance visual components
This commit includes significant refactoring of the frontend, specifically of the visual components and styling. The main changes include the addition of scrolling functionality, development of new table components including TableHeading, TableRow, TablePicture, and TableStages, as well as improvements and optimizations of existing UI components. Changes have been made to enhance consistency of layout in various files, as well as better visualization of data through the implementation of example data for vehicle issues. A new route "UI-components" has been introduced for more structured and efficient navigation.