Add 'Listeners' section and tooltip in ProblemListTable
A 'Listeners' column is added to the ProblemListTable to show all involved users. Each user's information is expanded with a tooltip that's shown on hover. Updated store types and data file to accommodate the new feature. Also, introduced a new width option for the Th component to correctly display the expanded user information.