Refactor problem list and enhance subscription functionality
Significant changes are made to the problem list interface and functionality. There has been a massive refactor in the problem list component, the introduction of the subscription/unsubscription feature, and several UI enhancements. A new component `SubscribeToListeners` was added that allows users to subscribe or unsubscribe from a problem, with the feature implemented in the store. The problem list table UI and structure have been heavily revised. The names of some fields (`userId` and `userName`) in the ProblemListTypes were changed to `name` and `designation`. An optional `defaultValue` and `clearable` prop were added to the `SelectSearch` component.