Enhance VehicleListsStore and add context hook
Updated the VehicleListsStore reducer to handle initial data and vehicles data, and also introduced a loading state. Created a new context hook, useVehicleListsContext, to provide easy access to VehicleListContext. This update also includes setting up initial skeleton for VehicleListHeader and addition of the loading state method in the VehicleListContext type.