Update dependencies in package-lock.json
Several dependencies were added, updated, and removed in package-lock.json. These modifications primarily focus on the frontend functionality, ensuring that the necessary dependencies are available and up-to-date for front-end development. This includes the addition of new packages like "@radix-ui/react-popover" and "date-fns" amongst others. Moreover, it includes the update to a different version of the "clsx" package and overall fine-tuning to improve the project's performance and stability.