Refactor problem registration UI and update documentation
The update enhances the UI in the NewProblemRegistration component by adding input components. The parameters in EditProblemInProblemList have also been adjusted accordingly. Moreover, the documentation files are revised to resolve a typographical error and is only built when there are changes on the 'docs' branch. These modifications are in line with the changes made in the problem registration feature.