Update UI components and implement barcode scanning functionality
Added 'framer-motion' for modal animations, extended the width of some columns in the ProblemListTable, and created a new ScanBarcode component. Also, updated dependencies and scripts in the package.json file and made changes to the UIComponents and VehicleSearch pages. Introduced logic for vehicle search header and fixed some modal functionalities.