Add color selection component to the admin project edit page
Two new components, ColorSelection and InputColor, have been added to the frontend. The former allows admins to select a color for a project in the AdminProjectEdit page, while the latter is a custom input field for picking a color. Further adjustments were also made to the layout and stage names input in AdminProjectEdit. Additionally, some minor changes were made to the AdminTypes and InputText components.