Add functionality to edit user permissions
New features added to the AdminUserList module enable editing of user permissions. This includes the ability to enable or disable permissions for accessing CFT Problem lists, displaying pictures on secured projects, and accessing the admin menu. The changes apply in real-time, and notifications are sent whenever a user's data is updated.