Add project organization chart in admin section
Introduced a new feature which includes an organization chart for projects in the admin section. Detailed data can now be listed and viewed by admins for each project, including managers and users with their respective roles. The addition of this chart will enhance project management and provide a clear overview of staff assignment.