Refactor Welcome page components and extend database setup
This commit involves refactoring the Welcome page components to enhance readability and code organisation. It also extends the database setup, adding missing entries and modifying existing ones to accommodate new requirements. A new file upload handler for the backend was added and a temporary logo redirection was fixed in the Navbar.